MILAB-4496 Minor fix + code clean up
